Fixed error due to package requirement issue
one of the packages required by cadquery was improperly configured and causes issues with the cadquery import.  Had to change docker base image and modify to include a configured environment to work around the issue.  May require rebuilding or removing and recreating your docker image and containers.

If you continue to encounter an import error please contact me.